feat: Refactor into classes, mostly working fix: Some refactoring so class setup is less reliant on JS
