+++ index.js
@@ -1,5 +1,5 @@
 if (typeof WebGLRenderingContext !== 'undefined') {
-  module.exports = require('./browser_index')
+  module.exports = require('./src/javascript/browser-index')
 } else {
-  module.exports = require('./node_index')
+  module.exports = require('./src/javascript/node-index')
 }
