+++ src/javascript/extensions/stackgl-destroy-context.js
@@ -1,0 +1,11 @@
+class STACKGLDestroyContext {
+  constructor (ctx) {
+    this.destroy = ctx.destroy.bind(ctx)
+  }
+}
+
+function getSTACKGLDestroyContext (ctx) {
+  return new STACKGLDestroyContext(ctx)
+}
+
+module.exports = { getSTACKGLDestroyContext, STACKGLDestroyContext }
