+++ src/javascript/extensions/stackgl-resize-drawing-buffer.js
@@ -1,0 +1,11 @@
+class STACKGLResizeDrawingBuffer {
+  constructor (ctx) {
+    this.resize = ctx.resize.bind(ctx)
+  }
+}
+
+function getSTACKGLResizeDrawingBuffer (ctx) {
+  return new STACKGLResizeDrawingBuffer(ctx)
+}
+
+module.exports = { getSTACKGLResizeDrawingBuffer, STACKGLResizeDrawingBuffer }
