+++ src/javascript/webgl-active-info.js
@@ -1,0 +1,8 @@
+class WebGLActiveInfo {
+  constructor (_) {
+    this.size = _.size
+    this.type = _.type
+    this.name = _.name
+  }
+}
+module.exports = { WebGLActiveInfo }
