+++ src/javascript/webgl-context-attributes.js
@@ -1,0 +1,22 @@
+class WebGLContextAttributes {
+  constructor (
+    alpha,
+    depth,
+    stencil,
+    antialias,
+    premultipliedAlpha,
+    preserveDrawingBuffer,
+    preferLowPowerToHighPerformance,
+    failIfMajorPerformanceCaveat) {
+    this.alpha = alpha
+    this.depth = depth
+    this.stencil = stencil
+    this.antialias = antialias
+    this.premultipliedAlpha = premultipliedAlpha
+    this.preserveDrawingBuffer = preserveDrawingBuffer
+    this.preferLowPowerToHighPerformance = preferLowPowerToHighPerformance
+    this.failIfMajorPerformanceCaveat = failIfMajorPerformanceCaveat
+  }
+}
+
+module.exports = { WebGLContextAttributes }
