+++ src/javascript/webgl-drawing-buffer-wrapper.js
@@ -1,0 +1,9 @@
+class WebGLDrawingBufferWrapper {
+  constructor (framebuffer, color, depthStencil) {
+    this._framebuffer = framebuffer
+    this._color = color
+    this._depthStencil = depthStencil
+  }
+}
+
+module.exports = { WebGLDrawingBufferWrapper }
