+++ src/javascript/webgl-shader-precision-format.js
@@ -1,0 +1,9 @@
+class WebGLShaderPrecisionFormat {
+  constructor (_) {
+    this.rangeMin = _.rangeMin
+    this.rangeMax = _.rangeMax
+    this.precision = _.precision
+  }
+}
+
+module.exports = { WebGLShaderPrecisionFormat }
